Move pattern code to abstract PatternBuilder
